silence cast warnings on 32-bit linux

+++ src/stasis/io/debug.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-/** @file 
+/** @file
 
   This implements a io handle that wraps another handle, logging each
   call (and its result) to stdout.
@@ -13,157 +13,157 @@ typedef struct debug_impl {
   stasis_handle_t * h;
 } debug_impl;
 
-static int debug_num_copies(stasis_handle_t * h) { 
+static int debug_num_copies(stasis_handle_t * h) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call num_copies(%lx)\n", (intptr_t)pthread_self(), (unsigned long)hh);  fflush(stdout);
-  int ret = hh->num_copies(hh); 
-  printf("tid=%9ld retn num_copies(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
+  printf("tid=%9ld call num_copies(%lx)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh);  fflush(stdout);
+  int ret = hh->num_copies(hh);
+  printf("tid=%9ld retn num_copies(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
-static int debug_num_copies_buffer(stasis_handle_t * h) { 
+static int debug_num_copies_buffer(stasis_handle_t * h) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call num_copies_buffer(%lx)\n", (intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
+  printf("tid=%9ld call num_copies_buffer(%lx)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
   int ret = hh->num_copies_buffer(hh);
-  printf("tid=%9ld retn num_copies_buffer(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout); 
+  printf("tid=%9ld retn num_copies_buffer(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
-static int debug_close(stasis_handle_t * h) { 
+static int debug_close(stasis_handle_t * h) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call close(%lx)\n", (intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
+  printf("tid=%9ld call close(%lx)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
   int ret = hh->close(hh);
-  printf("tid=%9ld retn close(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout); 
+  printf("tid=%9ld retn close(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   free(h->impl);
   free(h);
   return ret;
 }
-static lsn_t debug_start_position(stasis_handle_t *h) { 
+static lsn_t debug_start_position(stasis_handle_t *h) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call start_position(%lx)\n", (intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
+  printf("tid=%9ld call start_position(%lx)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
   int ret = hh->start_position(hh);
-  printf("tid=%9ld retn start_position(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout); 
+  printf("tid=%9ld retn start_position(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
-static lsn_t debug_end_position(stasis_handle_t *h) { 
+static lsn_t debug_end_position(stasis_handle_t *h) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call end_position(%lx)\n", (intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
+  printf("tid=%9ld call end_position(%lx)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
   int ret = hh->end_position(hh);
-  printf("tid=%9ld retn end_position(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout); 
+  printf("tid=%9ld retn end_position(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
-static stasis_write_buffer_t * debug_write_buffer(stasis_handle_t * h, 
+static stasis_write_buffer_t * debug_write_buffer(stasis_handle_t * h,
 						lsn_t off, lsn_t len) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call write_buffer(%lx, %lld, %lld)\n", 
-         (intptr_t)pthread_self(), (unsigned long)hh, off, len); fflush(stdout);
+  printf("tid=%9ld call write_buffer(%lx, %lld, %lld)\n",
+         (long)(intptr_t)pthread_self(), (unsigned long)hh, off, len); fflush(stdout);
   stasis_write_buffer_t * ret = hh->write_buffer(hh,off,len);
   stasis_write_buffer_t * retWrap = malloc(sizeof(stasis_write_buffer_t));
   *retWrap = *ret;
   retWrap->h = h;
   retWrap->impl = ret;
-  printf("tid=%9ld retn write_buffer(%lx, %lld, %lld) = %lx\n", 
-         (intptr_t)pthread_self(), (unsigned long)hh, off, len, (unsigned long)retWrap); fflush(stdout); 
-  return retWrap; 
+  printf("tid=%9ld retn write_buffer(%lx, %lld, %lld) = %lx\n",
+         (long)(intptr_t)pthread_self(), (unsigned long)hh, off, len, (unsigned long)retWrap); fflush(stdout);
+  return retWrap;
 }
-static stasis_write_buffer_t * debug_append_buffer(stasis_handle_t * h, 
-						   lsn_t len) { 
+static stasis_write_buffer_t * debug_append_buffer(stasis_handle_t * h,
+						   lsn_t len) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call append_buffer(%lx, %lld)\n", 
-         (intptr_t)pthread_self(), (unsigned long)hh, len); fflush(stdout);
+  printf("tid=%9ld call append_buffer(%lx, %lld)\n",
+         (long)(intptr_t)pthread_self(), (unsigned long)hh, len); fflush(stdout);
   stasis_write_buffer_t * ret = hh->append_buffer(hh,len);
   stasis_write_buffer_t * retWrap = malloc(sizeof(stasis_write_buffer_t));
   *retWrap = *ret;
   retWrap->h = h;
   retWrap->impl = ret;
-  printf("tid=%9ld retn append_buffer(%lx, %lld) = %lx (off=%lld)\n", 
-         (intptr_t)pthread_self(), (unsigned long)hh, len, (unsigned long)retWrap, ret->off); fflush(stdout); 
+  printf("tid=%9ld retn append_buffer(%lx, %lld) = %lx (off=%lld)\n",
+         (long)(intptr_t)pthread_self(), (unsigned long)hh, len, (unsigned long)retWrap, ret->off); fflush(stdout);
   return retWrap;
-  
+
 }
-static int debug_release_write_buffer(stasis_write_buffer_t * w_wrap) { 
+static int debug_release_write_buffer(stasis_write_buffer_t * w_wrap) {
   stasis_write_buffer_t * w = (stasis_write_buffer_t*)w_wrap->impl;
   stasis_handle_t * hh = w->h;
 
   // Debugging output adds a parameter, h, for convenience.
-  printf("tid=%9ld call release_write_buffer(%lx, %lx)\n", (intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)w_wrap); fflush(stdout);
+  printf("tid=%9ld call release_write_buffer(%lx, %lx)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)w_wrap); fflush(stdout);
   int ret = hh->release_write_buffer(w);
-  printf("tid=%9ld retn release_write_buffer(%lx, %lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)w_wrap, ret); fflush(stdout); 
+  printf("tid=%9ld retn release_write_buffer(%lx, %lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)w_wrap, ret); fflush(stdout);
   return ret;
 }
 static stasis_read_buffer_t * debug_read_buffer(stasis_handle_t * h,
-					      lsn_t off, lsn_t len) { 
+					      lsn_t off, lsn_t len) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call read_buffer(%lx, %lld, %lld)\n", 
-         (intptr_t)pthread_self(), (unsigned long)hh, off, len); fflush(stdout);
+  printf("tid=%9ld call read_buffer(%lx, %lld, %lld)\n",
+         (long)(intptr_t)pthread_self(), (unsigned long)hh, off, len); fflush(stdout);
   stasis_read_buffer_t * ret = hh->read_buffer(hh,off,len);
   stasis_read_buffer_t * retWrap = malloc(sizeof(stasis_read_buffer_t));
   *retWrap = *ret;
   retWrap->h = h;
   retWrap->impl = ret;
-  printf("tid=%9ld retn read_buffer(%lx, %lld, %lld) = %lx\n", 
-         (intptr_t)pthread_self(), (unsigned long)hh, off, len, (unsigned long)retWrap); fflush(stdout); 
-  return retWrap; 
-  
+  printf("tid=%9ld retn read_buffer(%lx, %lld, %lld) = %lx\n",
+         (long)(intptr_t)pthread_self(), (unsigned long)hh, off, len, (unsigned long)retWrap); fflush(stdout);
+  return retWrap;
+
 }
 
-static int debug_release_read_buffer(stasis_read_buffer_t * r_wrap) { 
+static int debug_release_read_buffer(stasis_read_buffer_t * r_wrap) {
   stasis_read_buffer_t * r = (stasis_read_buffer_t*)r_wrap->impl;
   stasis_handle_t * hh = r->h;
 
   // Debugging output adds a parameter, h, for convenience.
-  printf("tid=%9ld call release_read_buffer(%lx, %lx)\n", (intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)r_wrap); fflush(stdout);
+  printf("tid=%9ld call release_read_buffer(%lx, %lx)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)r_wrap); fflush(stdout);
   int ret = hh->release_read_buffer(r);
-  printf("tid=%9ld retn release_read_buffer(%lx, %lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)r_wrap, ret); fflush(stdout); 
+  printf("tid=%9ld retn release_read_buffer(%lx, %lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)r_wrap, ret); fflush(stdout);
   return ret;
 }
 
-static int debug_write(stasis_handle_t * h, lsn_t off, 
-		     const byte * dat, lsn_t len) { 
+static int debug_write(stasis_handle_t * h, lsn_t off,
+		     const byte * dat, lsn_t len) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call write(%lx, %lld, %lx, %lld)\n", (intptr_t)pthread_self(), (unsigned long)hh, off, (unsigned long)dat, len); fflush(stdout);
+  printf("tid=%9ld call write(%lx, %lld, %lx, %lld)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, off, (unsigned long)dat, len); fflush(stdout);
   int ret = hh->write(hh, off, dat, len);
-  printf("tid=%9ld retn write(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
+  printf("tid=%9ld retn write(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
-static int debug_append(stasis_handle_t * h, lsn_t * off, 
-		      const byte * dat, lsn_t len) { 
+static int debug_append(stasis_handle_t * h, lsn_t * off,
+		      const byte * dat, lsn_t len) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call append(%lx, ??, %lx, %lld)\n", (intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)dat, len); fflush(stdout);
+  printf("tid=%9ld call append(%lx, ??, %lx, %lld)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, (unsigned long)dat, len); fflush(stdout);
   lsn_t tmpOff;
-  if(!off) { 
+  if(!off) {
     off = &tmpOff;
   }
   int ret = hh->append(hh, off, dat, len);
-  printf("tid=%9ld retn append(%lx, %lld, %lx, %lld) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, *off, (unsigned long) dat, len, ret); fflush(stdout);
+  printf("tid=%9ld retn append(%lx, %lld, %lx, %lld) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, *off, (unsigned long) dat, len, ret); fflush(stdout);
   return ret;
 
 }
-static int debug_read(stasis_handle_t * h, 
-		    lsn_t off, byte * buf, lsn_t len) { 
+static int debug_read(stasis_handle_t * h,
+		    lsn_t off, byte * buf, lsn_t len) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call read(%lx, %lld, %lx, %lld)\n", (intptr_t)pthread_self(), (unsigned long)hh, off, (unsigned long)buf, len); fflush(stdout);
+  printf("tid=%9ld call read(%lx, %lld, %lx, %lld)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, off, (unsigned long)buf, len); fflush(stdout);
   int ret = hh->read(hh, off, buf, len);
-  printf("tid=%9ld retn read(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
+  printf("tid=%9ld retn read(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
 static int debug_force(stasis_handle_t *h) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call force(%lx)\n", (intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
+  printf("tid=%9ld call force(%lx)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh); fflush(stdout);
   int ret = hh->force(hh);
-  printf("tid=%9ld retn force(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
+  printf("tid=%9ld retn force(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
 static int debug_force_range(stasis_handle_t *h, lsn_t start, lsn_t stop) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call force(%lx,%lld,%lld)\n", (intptr_t)pthread_self(), (unsigned long)hh, start, stop); fflush(stdout);
+  printf("tid=%9ld call force(%lx,%lld,%lld)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, start, stop); fflush(stdout);
   int ret = hh->force_range(hh, start, stop);
-  printf("tid=%9ld retn force(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
+  printf("tid=%9ld retn force(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
-static int debug_truncate_start(stasis_handle_t * h, lsn_t new_start) { 
+static int debug_truncate_start(stasis_handle_t * h, lsn_t new_start) {
   stasis_handle_t * hh = ((debug_impl*)h->impl)->h;
-  printf("tid=%9ld call truncate_start(%lx, %lld)\n", (intptr_t)pthread_self(), (unsigned long)hh, new_start); fflush(stdout);
+  printf("tid=%9ld call truncate_start(%lx, %lld)\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, new_start); fflush(stdout);
   int ret = hh->truncate_start(hh, new_start);
-  printf("tid=%9ld retn truncate_start(%lx) = %d\n", (intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
+  printf("tid=%9ld retn truncate_start(%lx) = %d\n", (long)(intptr_t)pthread_self(), (unsigned long)hh, ret); fflush(stdout);
   return ret;
 }
 
@@ -188,7 +188,7 @@ struct stasis_handle_t debug_func = {
 };
 
 
-stasis_handle_t * stasis_handle(open_debug)(stasis_handle_t * h) { 
+stasis_handle_t * stasis_handle(open_debug)(stasis_handle_t * h) {
   stasis_handle_t * ret = malloc(sizeof(stasis_handle_t));
   *ret = debug_func;
   ret->impl = malloc(sizeof(debug_impl));
